tests: fix bf16 dtype mismatch in liger CE TP=1 grad assertion
torch.testing.assert_close checks dtype by default. The backward parity
assertion compared the Liger grad (cast to fp32) against the reference
grad (still bf16 when dtype=bf16), tripping a dtype-equality error
before the value comparison ran. Convert both grads to fp32 before the
assert.

Verified locally:
  python -m torch.distributed.run --nproc-per-node 1 -m pytest
    tests/unit_tests/fusions/test_liger_cross_entropy.py -k tp1
  -> 2 passed (bf16 + fp32)

  python -m torch.distributed.run --nproc-per-node 2 -m pytest
    tests/unit_tests/fusions/test_liger_cross_entropy.py -k tp2
  -> 2 passed (bf16 + fp32)

+++ tests/unit_tests/fusions/test_liger_cross_entropy.py
@@ -39,7 +39,7 @@ def test_liger_ce_matches_torch_cross_entropy_tp1(dtype):
     ).reshape(seq_len, batch)
     ref.sum().backward()
 
-    torch.testing.assert_close(grad_liger.float(), logits2.grad, **tols)
+    torch.testing.assert_close(grad_liger.float(), logits2.grad.float(), **tols)
     Utils.destroy_model_parallel()
